add openvswitch util func for getting dpid

+++ get-openvswitch.sh
@@ -1,9 +1,24 @@
 #!/bin/bash
 . utils.sh
-log installing rdo-release yum repo
-yum install -y https://www.rdoproject.org/repos/rdo-release.rpm
-log installing openvswitch 
-yum install -y openvswitch
-wget https://rawgit.com/alexfeigin/devel/master/openvswitch-selinux-policy-2.5.0-1.el7.centos.noarch.rpm
-yum install -y ./openvswitch-selinux-policy-2.5.0-1.el7.centos.noarch.rpm
-log Finished installing 
+if [[ -e /etc/profile.c/openvswitch.sh ]]; then
+	log openvswitch already installed
+else
+	log installing rdo-release yum repo
+	yum install -y https://www.rdoproject.org/repos/rdo-release.rpm
+	log installing openvswitch 
+	yum install -y openvswitch
+	wget https://rawgit.com/alexfeigin/devel/master/openvswitch-selinux-policy-2.5.0-1.el7.centos.noarch.rpm
+	yum install -y ./openvswitch-selinux-policy-2.5.0-1.el7.centos.noarch.rpm
+	log Finished installing 
+	cat <<-EOF > /etc/profile.d/openvswitch.sh
+	getdpid()
+	{
+		local br=\$1
+		if [[ "\$br" == "" ]]; then echo "usage getdpid <bridge-name>"; return; fi
+		local hexdpid=\$(ovs-vsctl -- get bridge \$br datapath-id 2>/dev/null)
+		if [[ "\$hexdpid" == "" ]]; then echo no bridge named \$br; return; fi
+		python -c "print(int(\$hexdpid,16))";
+	}
+	EOF
+	chmod +x /etc/profile.c/openvwitch.sh
+fi
